feat: SceneDelegate 삭제
