+++ Projects/App/Sources/AppDelegate.swift
@@ -34,23 +34,6 @@ final class AppDelegate: UIResponder, UIApplicationDelegate, MessagingDelegate {
     store.send(.appDelegate(.didFinishLunching))
     return true
   }
-  
-  func application(
-    _ application: UIApplication,
-    configurationForConnecting connectingSceneSession: UISceneSession,
-    options: UIScene.ConnectionOptions
-  ) -> UISceneConfiguration {
-    let configuration = UISceneConfiguration(
-      name: nil,
-      sessionRole: connectingSceneSession.role
-    )
-    
-    if connectingSceneSession.role == .windowApplication {
-      configuration.delegateClass = SceneDelegate.self
-    }
-    
-    return configuration
-  }
 }
 
 extension AppDelegate: UNUserNotificationCenterDelegate {
